Using parseDouble instead of parseLong to avoid crashing if a number contains a dot (issue #864)

+++ src/main/java/org/sigmah/server/servlet/exporter/utils/ExporterUtil.java
@@ -586,7 +586,7 @@ public final class ExporterUtil {
 						if (textAreaElement.getIsDecimal()) {
 							value = Double.parseDouble(strValue);
 						} else {
-							value = Long.parseLong(strValue);
+							value = (long) Double.parseDouble(strValue);
 						}
 						break;
 					case DATE:
